Update to bsnes v107r4 beta release.
byuu says:

  - bsnes: added video filters from bsnes v082
  - bsnes: added ZSNES snow effect option when games paused or unloaded
    (no, I'm not joking)
  - bsnes: added 7-zip support (LZMA 19.00 SDK)

[Recent higan WIPs have also mentioned bsnes changes, although the higan code
no longer includes the bsnes code. These changes include:

  - higan, bsnes: added EXLOROM, EXLOROM-RAM, EXHIROM mappings
  - higan, bsnes: focus the viewport after leaving fullscreen exclusive
    mode
  - bsnes: re-added mightymo's cheat code database
  - bsnes: improved make install rules for the game and cheat code
    databases
  - bsnes: delayed construction of hiro::Window objects to properly show
    bsnes window icons

- Ed.]

+++ bsnes/emulator/platform.hpp
@@ -16,7 +16,7 @@ struct Platform {
   virtual auto path(uint id) -> string { return ""; }
   virtual auto open(uint id, string name, vfs::file::mode mode, bool required = false) -> vfs::shared::file { return {}; }
   virtual auto load(uint id, string name, string type, vector<string> options = {}) -> Load { return {}; }
-  virtual auto videoFrame(const uint32* data, uint pitch, uint width, uint height) -> void {}
+  virtual auto videoFrame(const uint16* data, uint pitch, uint width, uint height) -> void {}
   virtual auto audioFrame(const double* samples, uint channels) -> void {}
   virtual auto inputPoll(uint port, uint device, uint input) -> int16 { return 0; }
   virtual auto inputRumble(uint port, uint device, uint input, bool enable) -> void {}
